refactor(folio): extract block measurement out of PagedEditor into core
Move the ~905-line measurement cluster (measureBlocks, measureBlock,
measureTableBlock, measureTableCellBlockVisualHeight, measureTextBoxBlock,
extractFloatingZones and helpers) from the 7.5k-line PagedEditor.tsx into
core/layout-engine/measure/measureBlocks.ts. These were always pure
module-level functions; core/layout-painter/renderPage.ts and
core/layout-bridge/{headerFooterLayout,footnoteLayout}.ts imported them back
from the React layer, a backwards core -> paged-editor dependency now removed.

Relocate layoutInstrumentation.ts into core/layout-engine/ so the measure
module can reach recordMeasureBlock as a sibling without re-introducing the
backwards dependency, keeping measureBlocks' signature unchanged.

Add measureBlocks.test.ts; the cluster is now unit-testable in isolation.
PagedEditor.tsx: 7536 -> 6631 lines.
